requirements.txt
streamlit==1.30.0
pillow==10.2.0
pytesseract==0.3.10
langchain==0.1.3
langchain-community==0.0.19
